Fix int overflow when checking image size

+++ components/insight/SRC/org/openmicroscopy/shoola/agents/imviewer/view/ImViewerModel.java
@@ -123,7 +123,7 @@ class ImViewerModel
 {
 
     /** Default maximum export size, 12kx12kx image */
-    static int DEFAULT_MAX_EXPORT_SIZE = 144000000;
+    static long DEFAULT_MAX_EXPORT_SIZE = 144000000;
     
 	/** The maximum size for the bird eye view for standard screen size.*/
 	private static final int BIRD_EYE_SIZE_LOWER = 128;
@@ -1367,13 +1367,13 @@ class ImViewerModel
         if (getPixelsData() == null)
             return false;
 
-        int imgSize = getPixelsData().getSizeX() * getPixelsData().getSizeY();
-        int maxSize = DEFAULT_MAX_EXPORT_SIZE;
+        long imgSize = (long)getPixelsData().getSizeX() * (long)getPixelsData().getSizeY();
+        long maxSize = DEFAULT_MAX_EXPORT_SIZE;
         String tmp = (String) ImViewerAgent.getRegistry().lookup(
                 LookupNames.MAX_EXPORT_SIZE);
         if (tmp != null) {
             try {
-                maxSize = Integer.parseInt(tmp);
+                maxSize = Long.parseLong(tmp);
             } catch (NumberFormatException e) {
                 ImViewerAgent
                         .getRegistry()
